requirements.txt
pytest>=3.12.3
selene>=2.0.0rc9
selenium>=4.32.0
allure-pytest>=2.13.5
python-dotenv~=1.1.0
undetected-chromedriver~=3.5.5
allure-python-commons~=2.14.2